Split up migration script
