+++ apps/services/search-indexer/src/migrate/lib/aws.ts
@@ -1,10 +1,10 @@
-import { logger } from '@island.is/logging'
 import AWS from 'aws-sdk'
 import { PutObjectRequest } from 'aws-sdk/clients/s3'
-import { environment } from '../environments/environment'
-import { Dictionary } from './dictionary'
 import { PackageStatus, DomainPackageStatus } from 'aws-sdk/clients/es'
 import { ElasticsearchIndexLocale } from '@island.is/content-search-index-manager'
+import { logger } from '@island.is/logging'
+import { environment } from '../../environments/environment'
+import { Dictionary } from './dictionary'
 
 AWS.config.update({ region: environment.awsRegion })
 const awsEs = new AWS.ES()
